Use the 'expiresIn' value, returned from the API, to set a timer to log the user out at the same time as their token expires
Committing as is to keep the commits small. However note that this commit has two flaws:

We don't set the time that the token expires to local storage, so if they leave the app before the timer expires, their token will still be stored in local storage, and we'll treat them as logged-in when the next come back (they just won't be able to do any actions on the database, which require them to be logged-in; in this case, change their password)

We don't clear the timer if they manually logout, so there will just be an errant timer counting down in the background.